refactor: @swagger 테스트 완료 - restcontroller

+++ spring-breeze-erp-test/back/src/main/java/com/sb/erp/resv/service/ReservationService.java
@@ -19,7 +19,7 @@ public interface ReservationService {
 
     int update(ResvRequest ResvDto);
 
-    int delete(int revId);
+    int delete(long revId);
 
     // 통계 (전체/승인/대기/반려)
     StatsResvResponse countByStats(ResvSearchRequest search);
